backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
SQLAlchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
celery==5.3.4
aiosmtplib==3.0.1
imapclient==3.0.1
python-decouple==3.8
httpx==0.25.2
motor==3.7.1
pydantic-settings==2.1.0      # <--- Omitir si solo usas decouple
python-dotenv==1.0.1          # <--- Opcional, para carga de .env
